Restructure file by taking redux folder outside the component folder to src folder and arranging it files respectively
