--- src/redux/Fakedat.js
@@ -1,19 +1,0 @@
-const BooksData = [
-  {
-    id: 1,
-    author: 'Suzanne Collins one',
-    title: 'yeap',
-  },
-  {
-    id: 2,
-    author: 'Suzanne Collins two',
-    title: 'yeap',
-  },
-  {
-    id: 3,
-    author: 'Suzanne Collins three',
-    title: 'yeap',
-  },
-];
-
-export default BooksData;
