+++ src/redux/category/Category.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import NavBar from './NavBar';
+import NavBar from '../books/NavBar';
 
 function Category() {
   return (
